kafka-test-python/pyproject.toml
[project]
name = "kafka-test-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.3.0",
]
